Move all fetch calls to apiCalls file
